% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cooks.distance.nls}
\alias{cooks.distance.nls}
\title{Cooks.distance for nls}
\usage{
\method{cooks.distance}{nls}(
  model,
  res = weighted.residuals.nls(model),
  sd = sqrt(stats::deviance(model)/df.residual(model)),
  hat = hatvalues.nls(model),
  ...
)
}
\arguments{
\item{model}{An object that inherits from class nls}

\item{res}{(possibly weighted) residuals, with proper default.}

\item{sd}{standard deviation to use, see default.}

\item{hat}{hat values \eqn{H_{ii}}{H[i,i]}, see default.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Cooks.distance for nls
}
\examples{
DNase1 <- subset(DNase, Run == 1)
## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
cooks.distance(fm1DNase1)
}
