% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-ci.R
\name{getbootci}
\alias{getbootci}
\title{Extract bootstrap confidence intervals}
\usage{
getbootci(boot.out, index, conf = 0.95)
}
\arguments{
\item{boot.out}{
    An object of class \code{"boot"} containing the output of a bootstrap
    calculation.  
  }

\item{index}{
    This should be a vector of length 1 or 2.  The first element of
    \code{index} indicates the position of the variable of interest in
    \code{boot.out$t0} and the relevant column in \code{boot.out$t}.  The
    second element indicates the position of the variance of the variable of
    interest.  If both \code{var.t0} and \code{var.t} are supplied then the
    second element of \code{index} (if present) is ignored.  The default is
    that the variable of interest is in position 1 and its variance is in
    position 2 (as long as there are 2 positions in \code{boot.out$t0}).
  }

\item{conf}{
    A scalar or vector containing the confidence level(s) of the required
    interval(s).
  }
}
\description{
Extract bootstrap confidence intervals
}
