% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hatvalues.nls}
\alias{hatvalues.nls}
\title{Approximation of hat values for nls.}
\usage{
\method{hatvalues}{nls}(model, ...)
}
\arguments{
\item{model}{An object that inherits from class nls}

\item{...}{further arguments for compatibility with S3 method.}
}
\value{
Vector of approximated hat values
}
\description{
Approximation of hat values for nls.
}
\details{
https://stackoverflow.com/questions/39167204/in-r-how-does-one-extract-the-hat-projection-influence-matrix-or-values-from-an
}
\examples{
DNase1 <- subset(DNase, Run == 1)
## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
hatvalues(fm1DNase1)
}
