\name{L.minor}
\alias{L.minor}
\docType{data}
\title{Enzyme kinetics}
\description{Enzyme kinetics}

\usage{data(L.minor)}

\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector}
    \item{\code{rate}}{a numeric vector}
  }
}

\source{
  Cedergreen, N. and Madsen, T. V. (2002) Nitrogen uptake by the floating macrophyte \emph{Lemna minor},
  \emph{New Phytologist}, \bold{155}, 285--292. 
}

\keyword{datasets}