% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlstools-package.R
\docType{package}
\name{nlstools-package}
\alias{nlstools}
\alias{nlstools-package}
\title{nlstools: Tools for Nonlinear Regression Analysis}
\description{
Several tools for assessing the quality of fit of a gaussian nonlinear model are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aursiber/nlstools}
  \item Report bugs at \url{https://github.com/aursiber/nlstools/issues}
}

}
\author{
\strong{Maintainer}: Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr}

Authors:
\itemize{
  \item Florent Baty \email{florent.baty@gmail.com}
  \item Marie-Laure Delignette-Muller \email{marielaure.delignettemuller@vetagro-sup.fr}
}

Other contributors:
\itemize{
  \item Sandrine Charles [contributor]
  \item Jean-Pierre Flandrois [contributor]
  \item Christian Ritz [contributor]
}

}
\keyword{internal}
