% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{O2K}
\alias{O2K}
\title{Oxygen kinetics during 6-minute walk test data set}
\format{
\code{O2K} is a data frame with 2 columns (t: time, VO2: oxygen
uptake)\cr
}
\source{
This data set was provided by the Cantonal Hospital St. Gallen,
Switzerland.
}
\usage{
O2K
}
\description{
Oxygen uptake kinetics during a 6-minute walking test in a patient with
pulmonary disease. The first 5.83 minutes correspond to the resting phase
prior to exercise.
}
\examples{

data(O2K)
plot(O2K)

}
\keyword{datasets}
