% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-ci.R
\name{pred.bs}
\alias{pred.bs}
\title{Function for bootstraping nls model and retrieve std errors}
\usage{
pred.bs(dataset, i, model, newdata)
}
\arguments{
\item{dataset}{original dataset}

\item{i}{vector of indices to bootstrap in dataset}

\item{model}{nls (or lm) model on which to do bootstrap}

\item{newdata}{new dataset for which to provide std errors}
}
\description{
Function for bootstraping nls model and retrieve std errors
}
\details{
http://r.789695.n4.nabble.com/standard-errors-for-predict-nls-td873670.html
}
