% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-ci.R
\name{predict_boot_nls}
\alias{predict_boot_nls}
\title{Confidence interval for predictions with nls}
\usage{
predict_boot_nls(object, newdata, R = 1000, conf = 0.95)
}
\arguments{
\item{object}{nls model on which to apply}

\item{newdata}{vector of new observation}

\item{R}{number of bootstrap to realise. Higher 
gives more precision but longer calculations}

\item{conf}{confidence interval to return}
}
\description{
Confidence interval for predictions with nls
}
\examples{
formulaExp <- as.formula(VO2 ~ (t <= 5.883) * VO2res + (t > 5.883) *
                           (VO2res + (VO2peak - VO2res) *
                              (1 - exp(-(t - 5.883) / mu))))
O2K.nls1 <- nls(formulaExp, start = list(VO2res = 400, VO2peak = 1600,
                                         mu = 1), data = O2K)
predict_boot_nls(O2K.nls1)
}
