% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{residuals_nls}
\alias{residuals_nls}
\title{Residuals of all types for nls models}
\usage{
residuals_nls(
  object,
  type = c("working", "response", "deviance", "pearson", "partial"),
  ...
)
}
\arguments{
\item{object}{An object that inherits from class nls}

\item{type}{the type of residuals which should be returned.
Can be abbreviated.}

\item{...}{further arguments for compatibility with S3 method.}
}
\description{
Residuals of all types for nls models
}
\details{
http://statweb.stanford.edu/~jtaylo/courses/stats306b/restricted/notebooks/quasilikelihood.pdf
}
\examples{
DNase1 <- subset(DNase, Run == 1)
## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
residuals_nls(fm1DNase1, type = "response")
residuals_nls(fm1DNase1, type = "pearson")
residuals_nls(fm1DNase1, type = "partial")
}
