% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vmkm}
\alias{vmkm}
\alias{vmkmki}
\alias{michaelisdata}
\title{Michaelis Menten data sets}
\format{
\code{vmkm} is a data frame with 2 columns (S: concentration of
substrat, v: reaction rate)\cr 
\code{vmkmki} is a data frame with 3 columns
(S: concentration of substrat, I: concentration of inhibitor, v: reaction
rate)
}
\source{
These datasets were provided by the French research unit INRA
UMR1233.
}
\usage{
vmkm
}
\description{
Michaelis Menten data sets
}
\examples{

data(vmkm)
data(vmkmki)
plot(vmkm)
plot(vmkmki)

}
\keyword{datasets}
