% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{weighted.residuals.nls}
\alias{weighted.residuals.nls}
\title{weighted.residuals for nls}
\usage{
weighted.residuals.nls(obj, drop0 = TRUE)
}
\arguments{
\item{obj}{An object that inherits from class nls}

\item{drop0}{logical.  If \code{TRUE}, drop all cases with
    \code{\link[stats]{weights} == 0}.}
}
\description{
weighted.residuals for nls
}
\examples{
DNase1 <- subset(DNase, Run == 1)
## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
weighted.residuals.nls(fm1DNase1)
}
