\name{nlstools-defunct}
\alias{nlstools-defunct}
\title{Defunct Functions in Package \pkg{nlstools}}

\description{
  The models or data sets listed here are no longer part of package \pkg{nlstools}. In order to access these models and data set in the future, please load the additional package \pkg{nlsMicrobio}.
}

\details{

Defunct functions are:\cr
\code{geeraerd}\cr
\code{geeraerd_without_Nres}\cr
\code{geeraerd_without_Sl}\cr
\code{mafart}\cr
\code{albert}\cr
\code{trilinear}\cr
\code{bilinear_without_Nres}\cr
\code{bilinear_without_Sl}\cr
\code{baranyi}\cr
\code{baranyi_without_Nmax}\cr
\code{baranyi_without_lag}\cr
\code{buchanan}\cr
\code{buchanan_without_Nmax}\cr
\code{buchanan_without_lag}\cr
\code{gompertzm}\cr
\code{jameson_buchanan}\cr
\code{jameson_baranyi}\cr
\code{jameson_without_lag}\cr
\code{cpm_T}\cr
\code{cpm_pH_4p}\cr
\code{cpm_pH_3p}\cr
\code{cpm_aw_3p}\cr
\code{cpm_aw_2p}\cr
\code{cpm_T_pH_aw}\cr
\code{competition1}\cr
\code{competition2}\cr
\code{growthcurve1}\cr
\code{growthcurve2}\cr
\code{growthcurve3}\cr
\code{growthcurve4}\cr
\code{ross}\cr
\code{survivalcurve1}\cr
\code{survivalcurve2}\cr
\code{survivalcurve3}\cr
}