\name{O2K}
\alias{O2K}
\docType{data}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{ Oxygen kinetics during 6-minute walk test data set }

\description{
Oxygen uptake kinetics during a 6-minute walking test in a patient with pulmonary disease. The first 5.83 minutes correspond to the resting phase prior to exercise.
}

\usage{
data(O2K)
}
%- maybe also 'usage' for other objects documented here.

\format{ 
\code{O2K} is a data frame with 2 columns (t: time, VO2: oxygen uptake)\cr
}

\source{ This data set was provided by the Cantonal Hospital St. Gallen, Switzerland. }

%\references{  }

\examples{
data(O2K)
plot(O2K)
}

\keyword{ datasets }% at least one, from doc/KEYWORDS
